Refactor useRampNavigation to remove providedQuoteId and streamline component rendering
